Sensitivity studies on the working dynamic model.
I have not been able to get close to reconciling all the dynamic
variables with their historical data series. I have not found a solution
so far that allows me to get more than one variable to match its data
series, and the fits for Y and w especially are sharply opposed. My
initial thought was that maybe Y would underestimate when the other
variables fit the data, as capital stock hasn't been included yet, but Y
overestimates in this case. This is the best fit so far, with a focus on
trying to line up the GDP data.